+++ doc/tools/gen_cmake_doc.py
@@ -14,6 +14,7 @@ DOCUMENTED_FILES=[
     "tests",
     "submodule",
     "option",
+    "codegen",
     "find",
     "flags",
     "swig/python"